consumption of supersaturation debugging

+++ src/PBE_general.f90
@@ -48,8 +48,8 @@ double precision :: amb_temp, amb_p, amb_rho, RH, part_den_l, alpha_ice, p_sat_l
 double precision :: jet_cl_model, diameter_jet, u_0j, T_0j, current_temp, current_rho, p_water, current_XH2O
 double precision :: tau_g
 double precision :: kappa !hygroscopicity
-double precision :: Loss_Sw !Saturation consumption
-double precision :: Smw !Saturation ratio along the mixingline with no particles
+real :: Loss_Sw !Saturation consumption
+real :: Smw !Saturation ratio along the mixingline with no particles
 
 integer m,grid_type
 integer i_gm,solver_pbe
